+++ pumpkin-chunk-system/src/chunk_system/channel.rs
@@ -1,5 +1,5 @@
 use super::{ChunkLevel, ChunkPos, HashMapType};
-use crate::chunk_system::chunk_state::StagedChunkEnum;
+use super::chunk_state::StagedChunkEnum;
 use crate::level::Level;
 use std::collections::hash_map::Entry;
 use std::mem::swap;
