+++ pumpkin-chunk-system/src/chunk_system/chunk_loading.rs
@@ -1,5 +1,5 @@
 use super::{ChunkLevel, ChunkPos, HashMapType, LevelChannel};
-use crate::chunk_system::chunk_state::StagedChunkEnum; // Fixed path
+use super::chunk_state::StagedChunkEnum;
 use itertools::Itertools;
 use std::cmp::{Ordering, PartialEq, min};
 use std::collections::BinaryHeap;
