+++ pumpkin-chunk-system/src/lighting/engine.rs
@@ -1,7 +1,7 @@
 use crate::chunk_system::Chunk;
 use crate::chunk_system::generation_cache::Cache;
-use crate::generation::height_limit::HeightLimitView;
-use crate::generation::proto_chunk::GenerationCache;
+use pumpkin_world::generation::height_limit::HeightLimitView;
+use pumpkin_world::generation::proto_chunk::GenerationCache;
 use crate::lighting::storage::{get_block_light, get_sky_light, set_block_light, set_sky_light};
 use pumpkin_config::lighting::LightingEngineConfig;
 use pumpkin_data::BlockDirection;
