+++ pumpkin-world/src/lib.rs
@@ -6,18 +6,16 @@ use pumpkin_util::math::vector2::Vector2;
 pub mod biome;
 pub mod block;
 pub mod chunk;
-pub mod chunk_system;
 pub mod cylindrical_chunk_iterator;
 pub mod data;
-pub mod dimension;
 pub mod generation;
 pub mod inventory;
-pub mod level;
-pub mod lighting;
 pub mod poi;
 pub mod tick;
 pub mod world;
 pub mod world_info;
+
+pub mod chunk_system_data;
 
 pub type BlockId = u16;
 pub type BlockStateId = u16;
